timeline demo
made a really basic demo of timeline re issue #7. A number of notes to
take

1) No CDN seems to host the timeline JS files. We will have to keep it
on our own server. I have included the 2 folders holding what we need
(css and js).

2)The data needs to come from a JSON file. working on generating this
from the MySQL server just now. For now I have a static one. Here is the
format we will need to conform to. Not just a straight up query.
https://github.com/NUKnightLab/TimelineJS#json

3)You HAVE to run it on a server. Will not work when ran locally, at
all. Simple thing stumped me for a good while.